% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatetab.R
\name{createdf}
\alias{createdf}
\title{Build a data.frame based on a format}
\usage{
createdf(name = "Design", nbrow = 100)
}
\arguments{
\item{name}{the sheet name}

\item{nbrow}{the number of row of the dataframe}
}
\value{
a data frame
}
\description{
Read sheet of an xlsx file containing a format definition and generate a data
frame
}
\author{
Laurent Dubroca
}
