% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{format_definition_fishframe}
\alias{format_definition_fishframe}
\title{Format definition for fishframe}
\format{A format definition in xlsx}
\description{
The format used by fishframe
}
\keyword{data}
