% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatetab.R
\name{pickvalue}
\alias{pickvalue}
\title{Pick values from a data frame to another if the names match}
\usage{
pickvalue(tab1, tab2)
}
\arguments{
\item{tab1}{the dataframe to which the values are copied}

\item{tab2}{the dataframe from which the values are coming from}
}
\value{
a data frame
}
\description{
Pick values from a data frame to another if the names match
}
\author{
Laurent Dubroca
}
