% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RDB_CS_Data_Model_v1.3}
\alias{RDB_CS_Data_Model_v1.3}
\title{Format definition for a new sampling format}
\format{A format definition in xlsx}
\description{
The format dev during WKSDECC and before
}
\keyword{data}
